Allow to inspect exit_code and other info on exceptions
DockerException classes format the relevant info nicely for user.

But sometimes code needs to use that info to take decisions or to digest that
for user.

For example, exit code of some `docker compose exec ...` command might be
relevant, and we would want to implement some logic that depends on exit code or
parse a json in stderr, etc.