Fix time argument formatting to include time zone
In local testing, I saw that `Container.logs` misbehaved when passing
the `since` parameter. I believe this is because my time zone is UTC-4,
and so asking for something like

    container.logs(since=container.state.started_at)

would return no logs as it was effectively asking for logs from the
future:
- container.state.started_at parses the inspect output into a zone-aware
  datetime; the timezone in inspect's output is always UTC (`Z`)
- the time formatting in `format_time_for_docker` strips the time zone
  rather than converts to local time
- the docker CLI treats its `--since` argument as being in local time

I instrumented the Popen call to see what arguments were passed to the
docker CLI. Taking the exact same command and appending a `Z` to the
time to force the correct time zone printed logs as expected.