Replace isort, black and flake8 with ruff (#617)
[Ruff](https://docs.astral.sh/ruff/) has emerged as the most popular
tool for linting, and covers the rules from flake8 among others. It also
serves as an (almost) drop in replacement for black and isort.

This PR replaces flake8, isort and black with ruff. 
* There are some minor differences is how black and ruff formats the
files, as seen in the files formatted in this PR
* Ruff with `F`, `E` and `W` covers the rules by the existing flake8
according to [these
docs](https://docs.astral.sh/ruff/faq/#how-does-ruff-compare-to-flake8)
* Updated pre-commit and github actions to also use Ruff