Fix handling of 'podman pod create --share=...' (#620)
Previously treated `Pod.create(share=[])` the same as not passing the
`share` arg (not passing it to podman), but the podman default is to
share `uts`, `ipc` and `net` namespaces. This means there's no way to
specify that *no* namespaces should be shared.

In addition, podman expects this list of namespaces to be passed as a
single comma-separated arg, e.g. `--share uts,ipc`, not in the form
`--share uts --share ipc` (only the last is taken in this case). This
means there's another bug where the use of the internal
`Command.add_args_iterable()` helper method meant only one shared
namespace could be specified.

It seems possible there could be other related bugs to do with the
format of lists on the CLI...