Display an error message if using the `python-on-whales` command (#640)
Before:

```
$ python-on-whales download-cli
zsh: command not found: python-on-whales
```

After:

```
$ python-on-whales download-cli
The python-on-whales command line interface has been removed in python-on-whales v0.74.0. If you were using it to download the docker CLI, you should instead do it manually. Take a look at https://gabrieldemarmiesse.github.io/python-on-whales/docker_client/ which has instructions to download the Docker client.
```