:bug: Fix buildx tests in the CI by upgrading the docker images (#646)
We test `buildx imagetools` by using docker images available on dockerhub, some are given hashes to get the image for a specific platform and thus checking that we parse the output correctly. The docker images used were old and thus we had to upgrade them. This is what I did in this PR.

Maybe we could find a way to avoid the dependency on dockerhub but I don't see any good solution for that.