Refactor `push` method to return CLI output when `stream_logs`
The `push` method now optionally returns an iterator with CLI
output when `stream_logs` is enabled. This output is a tuple
of the image name and the corresponding progress log line as
`bytes`. The `quiet` parameter must not be simultaneously
enabled as it conflicts. Updated `_push_single_tag` to
support this behavior. Currently only supports single-image
pushes.