Add service-ports flag to compose.run
Docker compose run supports --service-ports flag which, if specified
does (from the doc):

'''
The second difference is that the docker compose run command does
not create any of the ports specified in the service configuration.
This prevents port collisions with already-open ports.
If you do want the service’s ports to be created and mapped to the
host, specify the --service-ports
'''

This adds said flag to compose.run along with the necessary tests to
ensure the flag is added to the run invocation and that said invocation
does indeed create the port mapping (and only does so when the flag is
provided).

This will use Port 8074 during testing.